rename ticketmgr to mgr
