+++ go/app/ticket/internal/mgr/grpc.go
@@ -1,4 +1,4 @@
-package ticketmgr
+package mgr
 
 import (
 	"context"
